srcs/solidity/requirements.txt
asgiref==3.8.1
Django==4.2.13
sqlparse==0.5.0
typing_extensions==4.12.0
python-dotenv

djangorestframework==3.15.1
web3==6.1.0
